math: added Vector<2,3,4> operator== and operator!= res: added Image cli: added imagegen() to generate image resources from a *.jpg/png/bmp/tiff file and optional *.imageparams yaml: fixed Scalar caching always returning 0 (oops...) stb_image: compile only the loaders used by imagegen()

+++ oglTest/src/ui/drawable.cpp
@@ -8,11 +8,11 @@ namespace ui {
 struct pDrawable {
   Drawable::Type type;
 
-  virtual const Vertex *vertices() const   = 0;
-  virtual size_t numVertices() const = 0;
+  virtual const Vertex *vertices() const = 0;
+  virtual size_t numVertices() const     = 0;
 
-  virtual const u16 *indices() const      = 0;
-  virtual size_t numIndices() const = 0;
+  virtual const u16 *indices() const = 0;
+  virtual size_t numIndices() const  = 0;
 };
 
 struct pDrawableText : public pDrawable {
@@ -71,7 +71,7 @@ DrawableManager::DrawableManager() :
 
 unsigned DrawableManager::numAtlasPages()
 {
-  return m_local_atlas.size() / PageSize;
+  return (unsigned)m_local_atlas.size() / PageSize;
 }
 
 Color *DrawableManager::localAtlasData(uvec4 coords, unsigned page)
